+++ client/src/routes/Temples/hooks/useTempleNotificationReminder.ts
@@ -4,18 +4,18 @@ import {useTranslation} from 'react-i18next';
 import {Temple} from '../../../../../shared/src/types/Temple';
 import useExerciseById from '../../../lib/content/hooks/useExerciseById';
 import NS from '../../../lib/i18n/constants/namespaces';
-import useTriggerNotification from '../../../lib/notifications/hooks/useTriggerNotifications';
+import useTriggerNotification from '../../../lib/notifications/hooks/useTriggerNotification';
 
 const useTempleNotificationReminder = (
   temple: Temple,
 ): [Notification | undefined, (enabled: boolean) => Promise<string | void>] => {
-  const {name, contentId} = temple;
+  const {id, name, contentId} = temple;
 
   const {t} = useTranslation(NS.COMPONENT.SESSION_REMINDER);
   const exercise = useExerciseById(contentId);
 
   const [notification, setNotification, removeNotification] =
-    useTriggerNotification(temple.id);
+    useTriggerNotification(id);
 
   const setReminder = async (enabled = true) =>
     enabled
